refactor: id 변경

+++ src/main/java/org/sopt/post/controller/response/PostResponse.java
@@ -6,7 +6,7 @@ import org.sopt.post.domain.Post;
 import org.sopt.post.domain.constant.Tag;
 
 public record PostResponse(
-	int postId,
+	long postId,
 	String title,
 	String username,
 	String content,
